Handle in ActionHandler by block-by-block basis
